feat(api): break features to follow the domain
Breaking up the files allows us to better understand the different parts of the system
